auth_api/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
Babel==2.12.1
bcrypt==4.0.1
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.4
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.5.3
pluggy==1.0.0
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.8
PyJWT==2.7.0
pytest==7.3.1
pytest-asyncio==0.21.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
speaklater==1.3
SQLAlchemy==2.0.15
typing-extensions==4.6.2
unicorn==2.0.1.post1
urllib3==2.0.3
uvicorn==0.22.0
Werkzeug==2.3.4
WTForms==3.0.1
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
opentelemetry-api==1.18.0
opentelemetry-exporter-jaeger==1.18.0
opentelemetry-exporter-jaeger-proto-grpc==1.18.0
opentelemetry-exporter-jaeger-thrift==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-instrumentation-wsgi==0.39b0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
ua-parser==0.16.1
flask-swagger-ui==4.11.1
